+++ src/WindServiceProvider.php
@@ -29,11 +29,11 @@ class WindServiceProvider extends PluginServiceProvider
         if ($this->app->runningInConsole()) {
             $this->publishes([
                 __DIR__.'/../database/seeders' => database_path('seeders'),
-            ], 'wind-seeder');
+            ], 'zeus-wind-seeder');
 
             $this->publishes([
                 __DIR__.'/../database/factories' => database_path('factories'),
-            ], 'wind-factories');
+            ], 'zeus-wind-factories');
         }
 
         return parent::boot();