+++ src/Filament/Resources/LetterResource.php
@@ -22,8 +22,6 @@ class LetterResource extends Resource
 
     protected static ?int $navigationSort = 2;
 
-    protected static ?string $navigationGroup = 'Wind';
-
     protected static function getNavigationBadge(): ?string
     {
         return static::getModel()::where('status', config('zeus-wind.default_status'))->count();
@@ -34,35 +32,43 @@ class LetterResource extends Resource
         return $form
             ->schema([
                 TextInput::make('name')
+                    ->label(__('name'))
                     ->required()
                     ->disabled()
                     ->maxLength(255),
                 TextInput::make('email')
+                    ->label(__('email'))
                     ->email()
                     ->required()
                     ->disabled()
                     ->maxLength(255),
                 Select::make('department_id')
+                    ->label(__('department'))
                     ->options(Department::all()->pluck('name', 'id'))
                     ->required()
                     ->visible(fn (): bool => config('zeus-wind.enableDepartments')),
                 TextInput::make('status')
+                    ->label(__('status'))
                     ->required()
                     ->maxLength(255),
                 TextInput::make('title')
+                    ->label(__('title'))
                     ->required()
                     ->disabled()
                     ->maxLength(255)
                     ->columnSpan(['sm' => 2]),
                 Textarea::make('message')
+                    ->label(__('message'))
                     ->disabled()
                     ->maxLength(65535)
                     ->columnSpan(['sm' => 2]),
                 TextInput::make('reply_title')
+                    ->label(__('reply_title'))
                     ->required()
                     ->maxLength(255)
                     ->columnSpan(['sm' => 2]),
                 Textarea::make('reply_message')
+                    ->label(__('reply_message'))
                     ->required()
                     ->maxLength(65535)
                     ->columnSpan(['sm' => 2]),
@@ -73,11 +79,11 @@ class LetterResource extends Resource
     {
         return $table
             ->columns([
-                TextColumn::make('id')->sortable(),
-                ViewColumn::make('from')->view('zeus-wind::filament.message-from')->sortable(['name']),
-                TextColumn::make('title')->sortable(),
-                TextColumn::make('department.name')->sortable(),
-                TextColumn::make('status')->sortable(),
+                ViewColumn::make('from')->view('zeus-wind::filament.message-from')->sortable(['name'])->label(__('from')),
+                TextColumn::make('title')->sortable()->label(__('title')),
+                TextColumn::make('department.name')->sortable()->visible(fn (): bool => config('zeus-wind.enableDepartments'))->label(__('department')),
+                TextColumn::make('status')->sortable()->label(__('status'))
+                    ->formatStateUsing(fn (string $state): string => __("status_{$state}")),
             ])
             ->defaultSort('id', 'desc');
     }
@@ -88,5 +94,25 @@ class LetterResource extends Resource
             'index' => Pages\ListLetters::route('/'),
             'edit' => Pages\EditLetter::route('/{record}/edit'),
         ];
+    }
+
+    public static function getLabel() : string
+    {
+        return __('Letter');
+    }
+
+    public static function getPluralLabel() : string
+    {
+        return __('Letters');
+    }
+
+    protected static function getNavigationLabel() : string
+    {
+        return __('Letters');
+    }
+
+    protected static function getNavigationGroup() : ?string
+    {
+        return __('Wind');
     }
 }